fixed: reflection bug on OpenGL mode

+++ Assets/DeferredShading/Scripts/DSPECrawlingLight.cs
@@ -28,7 +28,7 @@ public class DSPECrawlingLight : DSEffectBase
     void UpdateRenderTargets()
     {
         Vector2 reso = GetDSRenderer().GetInternalResolution() * resolution_scale;
-        if (rtTemp[0] != null && rtTemp[0].width != reso.x)
+        if (rtTemp[0] != null && rtTemp[0].width != (int)reso.x)
         {
             for (int i = 0; i < rtTemp.Length; ++i)
             {
